Extracted the Draft Page onto its own component in preperation for the addition of routing
