+++ client/app/components/DraftPage/CardPreview/CardPreview.css
