+++ client/app/components/DraftPage/DraftPage.css
@@ -1,0 +1,20 @@
+.app-container {
+  display: grid;
+  grid-template-rows: 73% 27%;
+  top: 58px;
+  height: calc(100vh - 58px);
+  transition: grid-template-rows 1s;
+  -webkit-transition: grid-template-rows 1s;
+}
+
+.grid-without-stash {
+  grid-template-rows: 100% 0;
+  transition: grid-template-rows 1s;
+  -webkit-transition: grid-template-rows 1s;
+}
+
+#app-grid-top-row {
+  grid-row-start: 1;
+  display: grid;
+  grid-template-columns: 50% 50%;
+}
