+++ client/app/components/DraftPage/TheStash/StashItem.css
